Reconstruct and Implement backspace
