+++ include/shell.h
@@ -1,0 +1,7 @@
+#include "uart.h"
+#include "ftoa.h"
+
+void reset(int tick);
+void cancel_reset();
+int strcmp (const char *p1, const char *p2);
+void put_shell();
